Add files via upload
array functions has different array functions. It will create an array of a specified size with random numbers 0-7, values may be negative.
It will then write out the array to a pseudo-random filename

It has a function to check for diagonal dominance, as well as a function to read in from a file and initialize arrays for Jacobi iteration, as well as print 2d and print 1d array functions to test for correct information copying. 

The zip file contains matrices of various sizes.
I made some matrices not square. I couldn't remember if Jacobi iteration required a square matrix or not, so some values in the zip may be useless. I can create more as needed.