Add files via upload
These are almost identical. The only difference is in the randArray function. uppTriangle makes matrices in upper triangular format similar to gaussian eliminiation. Sparse and very sparse are similar. Sparse has about 50% of the matrix values as 0, but its up to random chance so it could be a bit more, could be a bit less.  and very sparse has 70% of the matrix values as 0.

ALL are still diagonally dominant